Use a Commit model in Javascript to parse commit data
This includes the prototype library in the bundle for convenience.
Whenever a new commit is selected, a new Commit object is created that
is going to parse the raw message. From the object we can select
author information, diffs, etc.

Currently there is still a bug with commits without content changes.