Change keyboard shortcuts for switching views
This remaps Command-1/2 to the History/Commit view.

The keys used to be used to switch the subview in the history view.
As these views are probably less frequently changed than the history/
commit view, we'd better use the easier shortcut for the latter.

The subviews can now be changed by using command-option-1/2/3