Index: Don't rearrange on every object
We used to change the PBFile status every time we found a 
similarly named file. Because the unstagedFilesController
had a filter and sort predicate for these, the array was
rearranged every time an existing file was found. This quickly
leads to stalling.

We fix it by temporarily disabling the automatic rearranging
and then reenabling it when we're finished.