GitRevList -- Try to read the encoding of the commit message
This tries to read the encoding of the commit message, if it's specified by
git (using i18n.commitEncoding at the time of committing). It uses an ugly
hack to try to convert the encoding string to an NSStringEncoding, and then
uses that for the subject and committer name.

I'm not sure how expensive this is. If it's too expensive, we might need to
cache this value.