Remove completely wrong TODO comment.
A deletion execution dialog should block the UI until it's complete if there could
potentionally be state change occurring which has different outcome for pre-deletion
or post-deletion finite state.

A middle ground solution would be to use a modal sheet attached to a specific document
which then only blocks this particular document.