Bug fix: Re-enable the toolbar item validation.
Although it is true that the remote can be chosen arbitrarily through the popup menus
we still need to have a specific local branch. All branches is not a specific branch.
Of course we could do away with determining the current local branch through what's 
selected in the branch popup menu and instead focus on what the appropriate controller's
current selection is.