Add preference options for refreshing the view + controller automatically on windowDidBecomeMain. Add preference for showing exact or relative dates. Make the showWhitespaceDifferences preference work. Make the context slider in the diff views work.
This is done with a minimal version of jQuery 1.3.2 being responsible for the animation and
a small cookie storing the last value between page refreshes so that it can be incremented by the HTML sliders
default value.
TODO: Maybe we can also make the default context value into a preference?