Bugfix: crasher on 10.5 due to webkit and the <button> element
Something in webkit doesn't like the button html element nor the -webkit-appearance: button css tag. Ends up crashing in one of various WebCore::ThemeMac::paint functions.

This would crash for anyone on 10.5 with the Gist it button enabled (which it is by default so that means anyone who ran it for the first time).

Changing the button back to the way it was before SHA:b557d6b5c6e675216d340d20833f39d0096b7ccc