Draw commit message line more accurately
When drawing the commit message line, the given rect should not be used
for anything other than clipping. According to the comment, the code that
draws the line relies on the text view always drawing complete lines.
However, it also draws fragments, specifically when the insertion point
is flashing. By ignoring the given rect, not only is our line more accurate
in its positioning, but we also handle the case where the insertion point
is sitting on top of the line.