Use libgit2 to filter working directory changes
For each changed path, discover the repository for that path and compare it to the repository that is open.  If they don't match, it's a submodule or another repo within the top-level one.  Either way, changes to that module shouldn't be committed to our repo (not that I offer any guarantee that GitX would do the expected commit correctly anyway…)

This fixes #1 - although I'm yet to test the performance on a big repo with many files/submodules.