Point new builds directly at the automatically-update app cast
For the time being, I'll assume that all users are interested in the development channel - especially as the 'stable' branch is significantly less stable.

Existing installs will point to the current build, which will still use the old URL when checking for updates.  This will continue until I manually refresh the .xml file at the old URL to point to a build with the new URL embedded.  I expect to do this sometime tomorrow.

Fixes #21