Preserve rectangle for drawing subject if refs too long
If the column for presenting refs is not wide enough to actually draw
any refs, the rectangle we started with that accounted for the history
lines already should not be adjusted with the rect(0,0,0,0).

In useful sized columns that zero rect would be updated as refs that can
be drawn are drawn ensuring the subject appears to the right.