Improve handling of multiple item selections
Displays a message in the diff area instead of the commit selected
before creating a multiple-commit selectction. And takes care of
disabling main/contextual menu items when multiple commits are selected.

We now have to pass a list of selected commits around rather than a
single commit as before, which means a number of classes had to be
modified.

Also tries to re-use the same logic for clipboard operations in the main
menu and contextual menus.