Move application activate for AppleScript diff commands to gitx command
As the application in AppleScript is that applications do not
automatically become active when performing an action, it appears
more consistent to not activate the application when opening a diff
and making the gitx command issue an activation command along with the
open diff command to keep the existing behaviour.

This also makes the implementation of open diff consistent with the
implementation of the open (document/repository) command.