Avoid exception when closing a repository window
It seems to be caused by PBGitHistoryList holding a weak reference
to PBGitRepository but also observing it via KVO which may lead to
the PBGitRepository object to be deallocated before the KVO-observer
is removed.

Attempt to solve this by implementing explicit setters in PBGitHistory
which also call the update actions that were previously triggered
by KVO.

If the atomicity of those fields was crucial before, this may require
a bit more work.