Fix display of renames (#46)
The option to show file moves in a single line
rather than a separate remove and add is not one
to set when creating the diff. It has to be used
to manipulate that diff in an additional step.

As all the parameters used here as NSString /
NSNumber-typed the compiler couldn’t detect
the option being passed to the wrong method :/

Also adapted the JS to display the file moving
again.

This behaviour was introduced in 7af30a7. It
should fix the issue described in #46.