De-deprecate those
The rationale is that I don't find the "standard" `-presentError:…` methods easy-to-use, because it expects the place where the error happened to always know whether the error should appear as a modal alert or a sheet, which is cumbersome.

So, for now, all our errors will go through that method, and get displayed as sheets.