Cleanup
While investigating #17, I found that code pretty hard to understand.

For the record, it built a set of OIDs while extracting the meaning of the `PBGitRevSpecifier`, which (since sets have no order) I suspected could cause the out-of-order behavior we had. Turns out the fix was much easier, but since I find this one cleaner, and less "loopy", here it is.