Fix a crash when having HEAD detached
With HEAD detached, the parseSymbolicReference code would find an OID as an unresolvedTarget, which would result in a nil branchRef, thus a nil _headOID, and a subsequent crash when trying to add this nil OID to a set "somewhere".

Simplify all that by just using whatever our underlying HEAD reference says.